Fragment Communication
1. We should not directly communicate with other fragments in a
fragment, if we do it in this way, we would make these two fragments
tightly coupled and would not be able to easily reuse these fragments.

2. Hence we need to use fragment communicator pattern, using which we
can easily communicate with other fragment and we would not be tightly
coupling these two fragments which are talking to each other
We can implement it by using a interface and implementing the interface
in the Activity which contain these fragments, the implemented method
would write to a fragment (Fragment 2)

3. Fragment 1 can use the implemented method in it, from dynamic
polymorphism from java. An interface reference variable can refer to a
subclass object, using this property we can give data to the
implemented method

4. That method which is implemented in the Activity can intern call the
Fragment 2’s method using fragment manager and  find view by id and
give the data
Which is a very Sophisticated way of communication between fragments